+++ e_metrobus/navigation/chart.py
@@ -1,5 +1,7 @@
 from collections import namedtuple
 
+import math
+from decimal import Decimal
 import plotly
 import plotly.graph_objects as go
 
@@ -10,7 +12,8 @@ MARGIN = 10
 OFFSET = 1
 SIZE = MARGIN - 2 * OFFSET
 TEXTSIZE = 5
-TROPHY_OFFSET = 10
+TROPHY_OFFSET = 2
+ICON_SIZE = 40
 
 DEFAULT_COLOR = "Gainsboro"
 FONT_COLOR = "#B0B0B0"
@@ -41,19 +44,32 @@ def get_sizes(max_value):
     return Sizes(*map(lambda x: x / 50 * max_value, [MARGIN, OFFSET, SIZE, TEXTSIZE]))
 
 
-def get_mobility_figure(values):
+def get_mobility_figure(values, title):
+    rounding = get_rounding(max(values))
+    if rounding == 0:
+        rounded_values = [int(v) for v in values]
+    else:
+        rounded_values = [round(v, rounding) for v in values]
     colors = [DEFAULT_COLOR] * 5
     colors[1] = E_BUS_COLOR
-    mobiles = [_("Zu Fuß/<br>Fahrrad"), _("E-Bus"), _("E-Pkw"), _("Dieselbus"), _("Pkw")]
-
-    max_value = max(values)
+    mobiles = [
+        _("Zu Fuß/<br>Fahrrad"),
+        _("E-Bus"),
+        _("E-Pkw"),
+        _("Dieselbus"),
+        _("Pkw"),
+    ]
+    scaled_values = [
+        (v + min(rounded_values)) / max(rounded_values) * 100 for v in rounded_values
+    ]
+    max_value = max(scaled_values)
     sizes = get_sizes(max_value)
 
     bar = go.Bar(
         x=mobiles,
-        y=values,
+        y=scaled_values,
         marker_color=colors,
-        text=values,
+        text=rounded_values,
         textposition="outside",
         width=0.6,
     )
@@ -75,7 +91,7 @@ def get_mobility_figure(values):
     fig.add_annotation(
         x=0.5,
         y=max_value,
-        text=_("CO<sub>2</sub> Emissionen [in g]<br>nach Verkehrsmittel"),
+        text=title,
         font={"size": 15, "color": FONT_COLOR},
         align="left",
         showarrow=False,
@@ -96,12 +112,15 @@ def get_mobility_figure(values):
             )
         )
     # Trophy Icons:
+    trophy_number = 1
     for i in range(3):
+        if i > 0 and scaled_values[i] > scaled_values[i - 1]:
+            trophy_number += 1
         fig.add_layout_image(
             go.layout.Image(
-                source=f"/static/images/icons/i_trophy_{i+1}_{'black' if i == 1 else 'gray'}.svg",
+                source=f"/static/images/icons/i_trophy_{trophy_number}_{'black' if i == 1 else 'gray'}.svg",
                 x=i,
-                y=values[i]
+                y=scaled_values[i]
                 + sizes.textsize
                 + sizes.size
                 + sizes.offset
@@ -111,6 +130,36 @@ def get_mobility_figure(values):
             )
         )
     fig.update_layout_images(
-        {"xref": "x", "yref": "y", "xanchor": "center", "yanchor": "top",}
+        {"xref": "x", "yref": "y", "xanchor": "center", "yanchor": "top"}
     )
     return DjangoFigure(fig, displayModeBar=False, staticPlot=True)
+
+
+def get_co2_figure(values):
+    title = _("CO<sub>2</sub> Emissionen [in g]<br>nach Verkehrsmittel")
+    return get_mobility_figure(values, title)
+
+
+def get_nitrogen_figure(values):
+    title = _("Stickoxid Emissionen [in g]<br>nach Verkehrsmittel")
+    return get_mobility_figure(values, title)
+
+
+def get_fine_dust_figure(values):
+    title = _("Feinstaub Emissionen [in g]<br>nach Verkehrsmittel")
+    return get_mobility_figure(values, title)
+
+
+def get_rounding(max_value):
+    if max_value >= 10:
+        return 0
+    str_value = str(max_value).split(".")
+    if len(str_value) == 1:
+        return 0
+    dec = str_value[1]
+    i = 0
+    while len(dec) > i and dec[i] == "0":
+        i += 1
+    if len(dec) == i + 1:
+        return i + 1
+    return i + 2